refactor: Rename figma handlers
